moved CLI functionality to a CLI module (cargo hack check fix)
